Move screenshot files to main.
