+++ projects/VisualStudio2013/mupen64plus-core.vcxproj
@@ -85,6 +85,7 @@
     <ClCompile Include="..\..\src\main\main.c" />
     <ClCompile Include="..\..\src\main\rom.c" />
     <ClCompile Include="..\..\src\main\savestates.c" />
+    <ClCompile Include="..\..\src\main\screenshot.c" />
     <ClCompile Include="..\..\src\main\sdl_key_converter.c" />
     <ClCompile Include="..\..\src\main\util.c" />
     <ClCompile Include="..\..\src\main\workqueue.c" />
@@ -112,7 +113,6 @@
     </ClCompile>
     <ClCompile Include="..\..\src\osal\files_win32.c" />
     <ClCompile Include="..\..\src\osd\osd.cpp" />
-    <ClCompile Include="..\..\src\osd\screenshot.c" />
     <ClCompile Include="..\..\src\device\rcp\pi\pi_controller.c" />
     <ClCompile Include="..\..\src\plugin\dummy_audio.c" />
     <ClCompile Include="..\..\src\plugin\dummy_input.c" />
@@ -302,6 +302,7 @@
     <ClInclude Include="..\..\src\main\main.h" />
     <ClInclude Include="..\..\src\main\rom.h" />
     <ClInclude Include="..\..\src\main\savestates.h" />
+    <ClInclude Include="..\..\src\main\screenshot.h" />
     <ClInclude Include="..\..\src\main\sdl_key_converter.h" />
     <ClInclude Include="..\..\src\main\util.h" />
     <ClInclude Include="..\..\src\main\version.h" />
@@ -311,7 +312,6 @@
     <ClInclude Include="..\..\src\osal\files.h" />
     <ClInclude Include="..\..\src\osal\preproc.h" />
     <ClInclude Include="..\..\src\osd\osd.h" />
-    <ClInclude Include="..\..\src\osd\screenshot.h" />
     <ClInclude Include="..\..\src\device\rcp\pi\pi_controller.h" />
     <ClInclude Include="..\..\src\plugin\dummy_audio.h" />
     <ClInclude Include="..\..\src\plugin\dummy_input.h" />
